// embedded-components/src/api/generated/ep-transactions-v3.schemas.ts
export type TargetCurrencyV3 = string;
// ...
export interface TransactionFxInformation {
  /**
   * Unique identifier for the foreign exchange rate
   * @minLength 1
   */
  rateId?: string;
  /**
   * The rate at which you charge for an FX transfer. You must express this as a decimal figure. For example, if you take a commission of 10%, the spread is expressed as 0.1
   * @minLength 1
   */
  clientSpread?: string;
}
// ...
export type RoutingInformationTransactionType =
  (typeof RoutingInformationTransactionType)[keyof typeof RoutingInformationTransactionType];

export const RoutingInformationTransactionType = {
  ACH: 'ACH',
  RTP: 'RTP',
  WIRE: 'WIRE',
} as const;
// ...
export type PaymentType = (typeof PaymentType)[keyof typeof PaymentType];

export const PaymentType = {
  RTP: 'RTP',
  WIRE: 'WIRE',
  ACH: 'ACH',
  TRANSFER: 'TRANSFER',
  CARD: 'CARD',
} as const;
// ...
export interface TransactionCounterPartyAddressDetails {
  /**
   * The type of the counterparty address. Valid values are:
| type | Description |
|-----------------|-------------|
| `RESIDENTIAL_ADDRESS` | Residential addresses |
| `BUSINESS_ADDRESS` | Business addresses |

   * @minLength 1
   */
  type?: string;
  buildingNumber?: BuildingNumber;
  /**
   * The street name of the counterparty address
   * @minLength 1
   * @maxLength 70
   */
  streetName: string;
  city: City;
  /** Two letter country code based on ISO 3166 standard, includes SEPA regions */
  country: string;
  countrySubDivision?: State;
  postalCode?: PostalCode;
}

export interface TransactionCounterPartyIdentityDetails {
  /**
   * Type of identification
   * @minLength 1
   */
  idType: string;
  /**
   * Identification value
   * @minLength 1
   * @maxLength 255
   */
  idValue: string;
}

/**
 * contact details of the counterparty
 */
export interface TransactionCounterPartyContactDetails {
  /**
   * The alternate name of the counterparty
   * @minLength 1
   */
  alternateName?: string;
  /**
   * The phone number of the counterparty
   * @minLength 1
   * @maxLength 30
   */
  phoneNumber?: string;
  /**
   * The email address of the counterparty
   * @minLength 1
   * @maxLength 256
   */
  emailAddress?: string;
  /**
   * The website of the counterparty
   * @minLength 1
   */
  website?: string;
}

/**
 * The party details of the counterparty
 */
export interface TransactionCounterPartyDetails {
  /**
   * The type of party this counterparty represents.
| Value | Description |
|-------|-------------|
| `INDIVIDUAL` | For individual party types |
| `ORGANIZATION` | For organization party types |

   * @minLength 1
   */
  type?: string;
  /**
   * The first name of the counterparty
   * @minLength 1
   * @maxLength 70
   * @pattern ^(.|\s)*\S(.|\s)*$
   */
  firstName?: string;
  /**
   * The last name of the counterparty
   * @minLength 1
   * @maxLength 70
   * @pattern ^(.|\s)*\S(.|\s)*$
   */
  lastName?: string;
  /**
   * The organization name of the counterparty
   * @minLength 1
   * @maxLength 140
   * @pattern ^(.|\s)*\S(.|\s)*$
   */
  businessName?: string;
  address?: TransactionCounterPartyAddressDetails;
  /**
   * The Identity details for the counterparty
   * @minItems 0
   */
  identities?: TransactionCounterPartyIdentityDetails[];
  contacts?: TransactionCounterPartyContactDetails;
}

/**
 * Routing number corresponding to the routing code type
 * @minLength 1
 * @maxLength 13
 * @pattern ^[A-Za-z0-9]([A-Za-z0-9\-]{0,11})[A-Za-z0-9]$
 */
export type RoutingNumber = string;

export interface TransactionCounterPartyAccountRoutingInformation {
  routingNumber: RoutingNumber;
  transactionType: RoutingInformationTransactionType;
  /**
   * Type of routing code.
   * @minLength 1
   */
  routingCodeType: string;
}
// ...
export interface TransactionCounterPartyExternalAccount {
  /**
   * The type of the counterparty account. Can be

| type | Description |
|-----------------|-------------|
| `CHECKING` | For checking accounts |
| `SAVINGS` | For savings accounts |
| `IBAN` | For IBANs |
| `NOT_APPLICABLE` | For when account type is not available |

   * @minLength 1
   */
  type: TransactionCounterPartyExternalAccountType;
  externalAccount: TransactionCounterPartyExternalAccountDetails;
}
// ...
export interface TransactionCounterPartyRegisteredAccount {
  /**
   * The type of the counterparty account. Should be `REGISTERED_ACCOUNT` for accounts that are registered on JPMorgan's system and can be looked up via the /accounts API.
   * @minLength 1
   */
  type: TransactionCounterPartyRegisteredAccountType;
  registeredAccount: TransactionCounterPartyRegisteredAccountDetails;
}
// ...
export interface TransactionCounterPartyCardAccount {
  /**
   * The type of the counterparty account. Should be `CARD` for card accounts that can be used for push to card transactions.
   * @minLength 1
   */
  type: TransactionCounterPartyCardAccountType;
  card: TransactionCounterPartyCardAccountDetails;
}

/**
 * The account details of the counterparty
 */
export type TransactionCounterPartyAccount =
  | TransactionCounterPartyExternalAccount
  | TransactionCounterPartyRegisteredAccount
  | TransactionCounterPartyCardAccount;

export interface TransactionCounterParty {
  /**
   * The id of the counterparty
   * @minLength 1
   * @maxLength 36
   */
  id?: string;
  /**
   * The type of the counterparty
| Value | Description |
|-------|-------------|
| `ON_BEHALF_OF` | When performing TP3 (Third Party) Flows this field needs to be set |

   * @minLength 1
   */
  type?: string;
  partyDetails?: TransactionCounterPartyDetails;
  account?: TransactionCounterPartyAccount;
}

/**
 * The purpose of the payment
 */
export interface TransactionPaymentPurpose {
  /**
   * Purpose of the payment in coded form. Must be between 1 and 4 characters.
   * @minLength 1
   * @maxLength 4
   */
  code?: string;
  /**
   * Purpose of the payment in proprietary form. Must be between 1 and 35 characters.
   * @minLength 1
   * @maxLength 35
   */
  customCode?: string;
}

export interface PostTransactionRequestBaseV3 {
  /**
   * A unique reference ID that you must generate and provide for each transaction.
   * @minLength 1
   * @maxLength 35
   * @pattern [_0-9A-Za-z]+
   */
  transactionReferenceId: string;
  type: PaymentType;
  /**
   * The amount that is moved out or into your account in the currency specified
   * @minLength 1
   * @maxLength 40
   * @pattern ^\d+(\.\d+)?$
   */
  amount?: string;
  /**
   * This denotes the currency of the payment in ISO 4217 Alpha-3 Currency Code format
Supported values 
| Code | Currency |
|------|----------|
| USD | United States Dollar |

   * @minLength 3
   * @maxLength 3
   */
  currency: string;
  targetCurrency?: TargetCurrencyV3;
  /**
   * Each ACH payment uses a Standard Entry class code (SEC) to indicate account type and authorization method used. <br>
Defaults to PPD for consumer, `CCD` for business, and `WEB` for direct debits ACH if no mandate method is specified. <br>
See documentation for further details <br>

| LocalInstrumentCode | Meaning                                                                |
|---------------------|------------------------------------------------------------------------|
| * `CCD`                 | Urgent Payment Net Settlement Corporate Credit or Debit (CCD)          |
| * `PPD`                  | Pre-arranged Payments or Deposit Transfer(PPD)                         |
| * `TEL`                  | TELEPHONE Initiated Entry - Used for authorizing payment via telephone |
| * `WEB`                  | WEB Initiated Entry - Used for authorizing payment over the internet   |

   * @minLength 1
   * @maxLength 35
   */
  localInstrumentCode?: string;
  /**
   * Transaction information that states its purpose for easier tracking
   * @minLength 1
   * @maxLength 140
   * @pattern ^(.|\s)*\S(.|\s)*$
   */
  memo?: string;
  fxInformation?: TransactionFxInformation;
  debtor?: TransactionCounterParty;
  creditor?: TransactionCounterParty;
  paymentPurpose?: TransactionPaymentPurpose;
}